undertaker-checkpatch: analyze patchfiles
This change introduces a new python module for analyzing patchfiles.
The basic functionality is to apply the patch to a Linux tree and check
if any defects are introduced, fixed, changed, or if they remained
present (unchanged).

In addition, undertaker-checkpatch can further analyze a defect's cause
and effect (-c/--check):

- Missing defects (symbolic) are caused by (Kconfig) identifiers that
    are referenced in the source code but are absent in Kconfig
    (i.e., not defined in any Kconfig file). undertaker-checkpatch
    finds such identifiers and prints them on stdout.

- Code defects (logic) are caused by conflicts in the CPP structure of
    source files. First, they can be caused by CPP items that are
    previously set (#define/#undefine CONFIG_XXX) in the source code.
    Such cases are identified automatically and printed on stdout.
    Second, code defects may also be caused by contradictions in the
    logical structure of #ifdef blocks (e.g., #ifdef CONFIG_X86 ...
    #ifndef CONFIG_X86). In this   case, undertaker-checkpatch prints
    the CPP structure of affected blocks to simplify manual analysis.

- Kconfig defects (logic) occur due to logical conflicts of Kconfig
    constraints in the precondition of an #ifdef block. Hence, we
    cannot automatically identify the causing Kconfig features, so
    that checkpatch generates a minimally unsatisfiable subformula
    (-u/--mus) to simplify manual analysis.

Furthermore, undertaker-checkpatch parses changes to Kconfig files and
reports if they break the referential integrity in the source. For
instance, if a Kconfig feature is removed or renamed but still
referenced in some files. Same, if a patch adds Kconfig statements
(if, depends on, select) which reference undefined features.

Change-Id: I723286c3b5b9b99b364ecc06fc5b6e9456902c38