undertaker: introduce incremental SatChecker
We created a new SatChecker-Object for every formula we wanted to
analyze. For dead-analyses we could reuse SatChecker-Objects and
reduce the reallocating overhead of picosat. This results in a
performance increase.
CnfModels can now be loaded directly when a SatChecker-Object is
created instead loading it indirectly over doIntersect (adding
'._.$ARCH._.' to the formula, search the string with a regex and load
the identified model).
The Satchecker can be incremented with 'operator(new_formula)'. The
call to 'operator(new_formula)' also checks if the merged formula
(old + new) is satisfiable.

Change-Id: I824cdcd3d815a06d724b92016bed81e2b19e5061