undertaker: remove unused debug remains / pprinter in SatChecker class
Remove old and unused debugging functions and variables from the
SatChecker class.
With these functions removed, it came to attention that the pprinter
functionality was also absolete and harmful for performance. The
SatChecker::pprinter printed only the given sat-formula and called the
SatChecker on this formula without doing anything with the result.
In an older version of the satchecker, this call was necessary to
print a debug formula for the old version of the parser.

Change-Id: I7cf485e6b465002f86dd283a5c05ea2074c7c78f