golem/inference_atoms: exit when no model is present for architecture
When inferences are calculated without a model, the symbol information
is incomplete and the resulting formulae are sometimes wrong (i.e.
when a symbol is only allowed to be selected as module or disabled).
These wrong formulae can lead to wrong results, therefore we choose
the conservative approach now and exit when no model is present for
the analyzed architecture.

Change-Id: Ia51c327ac97cb5606f64ab1be505e6c54064dda0