tailor: Use UniqueStringJoiner instead of StringJoiner for formula
So far, preconditions for traced blocks were simply aggregated and put
into the StringJoiner object sj.

As quite a lot of block preconditions are equal (i.e., have the same
formula), the resulting overall formula describing the whole scenario
has some redundant sub-formulae.

Using a UniqueStringJoiner for aggregation reduces the size of the big
formula (and thus, time required to solve it) significantly.

Change-Id: I518b136a9111b11b2562ddc49059f79e8d48ae71