scripts/kconfig: don't abort when 'mainmenu' isn't the first entry
When a 'mainmenu' entry is not the first entry in a toplevel
Kconfig-file, our version of kconfig aborts with an error message
'invalid option'.
While this is not a problem in newer linux versions, older versions
(i.e. v2.6.31 with arch frv) have 'mainmenu' entries which aren't the
first entry in a toplevel Kconfig-file. That fact lead dumpconf to
create empty .rsf-files and therefore emptry models for these old
versions of linux.

Change-Id: Ifd6d4aeec015e4e9881d91fa697d0579baaff14b