undertaker: rewrite help messages
The interface of the undertaker is comparatively complex. Some help
message don't really help in understanding the functionality or don't
explain the output format (e.g., -j cppsym).

I consider a complex interface in combination with partially poor help
messages to be a serious usability issue. This change is supposed
to trigger rewriting the help messages and, if needed, changes to the
interface as well.

Some examples:

The jobs are listed last, but aren't they the most important options?
If I want to check the functionality of a tool, then I primarily want
to find the main functions first and don't care about lists or a
'minimally unsatisfiable subformula'.

- cppsym: statistics over CPP symbols mentioned in source files
What to expect? The output did not explain at all that the numbers
represent.

- model:    print all options which are in the configuration space
What's the configuration space? Are people outside the project supposed
to know what this is?

- cpp: print on stdout cpp -D command line arguments
- exec:cmd: pipe file for every configuration to cmd
- model:    print all options which are in the configuration space
- all:      dump every assigned symbol (both items and code blocks)
- combined: create files for both configuration and pre-commended sources
The alignment is really ugly and not consistent.

Removed '-O exec:cmd:' from the helptext, since it's only used for
testing the formatCommented option and didn't indiate at all that the
output is representing the formatCommented-Option.

Removed 'interactive mode' from the helptext, since its functionality
is pretty limited right now, not really useful and the text didn't
indicate how interactive mode can be used.

Reported-by: Valentin Rothberg <valentinrothberg@gmail.com>

Change-Id: I50008c737a4619419b8c3d61e1c9ede135e7580a