BoolExpLexer/Parser: don't unnecessarily re-create parser files
We are currently using flex/bison to create the BoolExpLexer /
BoolExpParser source-files whenever they aren't present. Therefore,
all users of the undertaker have to have flex and bison installed.
However, the generated code doesn't vary that much.
I used the Kconfig-parser as example for a buildsystem without
implicit recreation of the parser source-files.
The creation of the BoolExp-Parser source-files is now an explicit
target in the Makefiles. This target needs a variable to be set to be
available.
'make REGENERATE_PARSERS=1 regenerate_parsers'

This removes the dependencies for flex and bison.

Change-Id: I58be58bc642cbfd9405442aeb81b4dfbae15c477