undertaker: correctly filter targets for .d file inclusion
So far, the findstring function of make was used to identify the
targets for which the .d files do not need to be included. This was
intended to have make only generate the files if needed (e.g., not
generate them when cleaning).

Using findstring to identify them however is problematic, as it
also matches substrings. This leads to "make check" matching
"clean-check" in $(NODEPS), which in turn leads to the .d files not
being included, breaking the build process before checking.

The intended behaviour is reached by using the filter function instead.

Change-Id: I169d8398b96b9e5a823c7a7fc20c8aa85a43566b