fm: fix dependencies for parallel checking
This change resolves a problem that occurs when using "make check"
with multiple threads.

The targets in $(ALLCONFIGS) all depend on $(CONF), which has not been
explicitely modelled yet. When executing "make check", the prerequisites could
be evaluated in parallel. Due to a missing $(CONF) binary evaluating
$(ALLCONFIGS) will then fail.

With this change, $(CONF) and $(SATYR) are added as a prerequisite for the
testcases, and removed from the prerequisites of the check target.

Change-Id: I8a177af5cd0a721bec9b2f05271f31657bec5d7a