undertaker: fix parallel building
The parallel build process became racy on some machines. The stack.hh
header of the BoolExpParser files was only mentioned in the .d file
and not in the makefile, therefore it was created later than the other
parser-header files but the build system tried to compile the parser
before it was copied. This lead to a build-failure on some machines.

Change-Id: I2ad665c822386c41790900ac88288223cce76e4f