Fix parallel compilation in the undertaker/ subdirectory
When compiling the undertaker/ directory with multiple threads,
object files are compiled multiple times, as the sub-makes (started
by the single targets undertaker/undertaker, undertaker/satys etc. in
the top level Makefile) are not aware of each other.

This change will make the undertaker/ directory only be entered once
but still parallel to the building of the other programs.

Change-Id: I59299e33905576eb811c426d4161841ad4cae9d2