buildsystem: add target 'localpuma' to download and build puma
To be able to easily access the undertaker on non-debian based
distributions, a new target 'make localpuma' is added.
This target will download the pre-woven aspectc++ and Puma sources and
trigger the compilation of the undertaker and Puma.

Additionally, an error is triggered when a test-compilation with
-lPuma isn't sucessfull and neither LOCALPUMA is defined nor
'localpuma' is the current target.

Change-Id: I2a182e002db96ce5d28a526b67458dbbfb37950e