undertaker-checkpatch: check all RSF models
In case a patch adds identifiers to Kconfig expressions
or statements, undertaker-checkpatch checks if such
identifiers are actually defined in Kconfig. This is
accomplished by checking the corresponding RSF files
in the models directory.

Until now, if no arch is specified, only the x86 model
is checked, what leads to some false positives. This patch
changes the process to check _all_ RSF models in the
models directory.

In case a Kconfig file in the HAL (e.g., arch/x86/Kconfig)
is changed, only the affected arch-specific model will be
checked. This avoids certain false negatives where a new
reference on identifier A is not defined in the affected
architecture but in others.

I found this bug by analyzing the following Linux
Git commit:
031d705c07bc7b911c49c49ae346e3223b6ced9b

Change-Id: I851ae4e59e895f3f4e6ee92b85a8a5abae00918c