undertaker-checkpach: arch-specific model generation
We can specify an architecture as well as a path to models for
undertaker-checkpatch. This allows us to check only one architecture
or just a specified set of models and thereby improves the overall
run time of the tool.

However, these options are not exclusive and may lead to the case
that the desired architecture is not present in the specified models
path. This patch adds a check for this case and generates a model
for this architecture if needed.

Change-Id: I27765b5b3c421fc6d1d484c7c847ea572619bb6f