undertaker-kconfigdump: input validation for manually specified arches
When undertaker-kconfigdump was given a list of architectures to
process, the exit-status of the script was still 0, even if one or
more of these architectures didn't exist.
undertaker-kconfigdump now checks if all corresponding Kconfig-Files
are present and exits with an status representing the number of absent
files if there are any.

Change-Id: I9e32786ea99eb8aab09c0163db8adaa1bcbe403a