undertaker: fix linking for testprograms when run-lcov is executed
Since the compilation of the test-programs was removed from 'make all'
in the undertaker/Makefile, the LDCOV flag (and with that the
necessary linking flags for lcov) wasn't present when the
test-programs were compiled and linking failed.

This error wasn't noticed because the call in run-lcov is '-$(MAKE) check'.
A leading '-' ignores errors.

Change-Id: I91de2fc290cf7cb5ae02e5d244e345cd81a9ad9e