undertaker-checkpatch: exception handling for model generation
This patch adds a proper exception handling for the generation of
variability models. Now, the exceptions and errors of
tools.generate_models() are caught in undertaker-checkpatch, which
exits with the corresponding error messages.

There are several error cases for the model generation, such as
the absence of a desired architecture ...

Cannot generate models:
Generating Format 1.0 (rsf) models
ERROR: Architecture TEST_ARCH not found

... or OS errors like bad file system permissions ...

Cannot generate models:
[Errno 13] Permission denied: '/home/valentin/linux/models/'

Change-Id: I3ed86d591bd4405446d3946329e4d3d95cdf6f97