undertaker-checkpatch: speed up loading of RSF models
The constructor of RsfModel parses the model file and reads
the RSF file afterwards. Reading an RSF file is very expensive and
not needed for checkpatch, which only needs to know if certain
identifiers are defined or not; this task is done by parsing the
model in the first step.

This commit changes the interface of the RSF model and adds a new
parameter readrsf (default=True) which indicates if the RSF file
should be read or not. Now, undertaker-checkpatch sets
readrsf=False and thereby speed ups the loading of RSF models.

I experienced a great loss in performance after merging commit
cec10802b1a0676538712e1917c478722c936a76 as checkpatch now loads
all models in the model directory and not just one. With this
change, the runtime of only loading all models of Linux v3.17 is
reduced from 0m39.753s to 0m0.733s on my machine.

Change-Id: I9a6b047296390b471bbb9e204474b44f416d43f3