defect_analysis.py: don't raise Exception for dead analysis
The undertaker fails for some files, mostly such that contain
assembly code. The initial intention was to ignore the return
code of the undertaker and collect all report informations
from stdout ('^I:'). However, the defect analysis rose a
CommandFailed exception until now. This commit changes the
behavior to the desired ignoring of the return value.

Change-Id: I139c50e1ffb1c66159941f6ccf608ff3676de1b9