undertaker: initiate all pointer members to 'nullptr'
To prevent bugs with uninitialized pointers as members in classes, we
now initialize all member-pointers with 'nullptr'. This will result in
no overhead when the member is initialized in the constructor (the
compiler won't create code for the nullptr-initialization in that
case) and if we change the behaviour of constructors, we don't risk
uninitialized pointers with bogus behaviour.

Change-Id: I638dbdfd209c8c4596cbde46d9699db8651a0c67