RsfReader: Add ItemRsfReader as "shallow" variant of RsfReader
Using shlex to parse the RSF model files takes quite a while to
complete. Additionally, the lookup of the symbol type is in O(n)
because an iteration over the whole database is performed.

Most of the time, however, not all information from the RSF file
is actually required and lookup should be fast.

This change adds a class ItemRsfReader which only saves information
about the item type and uses the dict to look up the type.

Change-Id: I8d9e4f018e83626bc8fc799e0080bec4af7c7cc9