validation: test-undertaker-checkpatch
This change adds a script which does some black box testing of
undertaker-checkpatch.

Checkpatch is particularly hard to integrate into the test suit,
since it can only be executed within a Linux Git. Consequently,
this test script checks various Git commits that I identified to
cause variability defects and referential violations with varying
parameters, trying to execute as much of checkpatch's code as
possible. Note that this script must be executed in a Linux Git.

The execution of the script takes 14 minutes on my machine with
Core(TM) i7-4600U, 16 GB of RAM and a solid state hard drive.

Change-Id: I3ff32a0cb594bd291bdfd84c4cac5c5f948fe6f8