python/model: Rewrite CnfModel class, ensure consistent behaviour
When parsing a CNF model, regular expressions were used
to match a simple pattern. This can be rewritten to use
.startswith() + .split() instead, which is faster.

Additionally, parsing can be stopped once a line starting
with "p cnf" is found as only the formula comes after that.
These two changes reduce the time to load significantly
(0.8 to 0.03 seconds on my machine).

Furthermore, up to now get_type() and is_bool_tristate()
in CnfModel caused a RuntimeError when the symbol was not
found in the model. In an RsfModel, these functions are
forwarded to RsfReader which will simply return None/False.

This changes CnfModel to also return None (get_type()) or
False (is_bool_tristate()) if the symbol is not present
in the model.

Change-Id: Id753299bb649ba47ddfd3857e28278eb64180ed7