undertaker/SatChecker: move MUS-functionality to SatChecker class
We often feed very long formulae into the satchecker. When we want
to retrace which symbols are causing a defect, it is almost impossible
to do that because of the length of the formula.
Therefore it is reasonable to move the MUS functionality into the
SatChecker-class, to be able to get a Minimized Unsatisfiable Subset
out of every SatChecker object.

Change-Id: I42d0fad8bfa630b0a9136bb587b3b1f56472ab98