undertaker-checkpatch: Fix classification of defects
After the last restructuring change (commit 87383f00)
defect_analysis.compare_blocks() did not cover all cases that could
occur during the classification of two defects.

If we found a mapping between two blocks (block_a and block_b)
and block_a had a defect while block_b did not, the control
flow reached the case for a "Changed defect" and would output
...XX.no_defect as the "target" defect.

Now, this is tested earlier and the correct classification as a
"repaired" defect is triggered. Comments have been added to clarify
the current 'decision state' we're in while going through the if
cascade.

Additionally, the requirement for block.match to be 'None' was dropped
in the main() function when printing 'repaired' defect reports. A
matching block could still be present while not being dead or undead
anymore.

Change-Id: I4b5e4ff78c9a6ae5d40dc242da10e0ce1205bd13