python: Make TranslatedModel.__unicode__() fast
Currently, TranslatedModel.__unicode__() uses a string to accumulate
the translated model in. Lines are added by doing '+=' on that string.

The problem with this is that strings are immutable and consequently
'+=' resolves into the creation of a new string and overwriting the
current reference with the new, extended one.

This patch replaces '+=' with a list, .append() operations and a
final ''.join(). This speeds up model generation on my machine from
~23 seconds to 4 seconds.

Change-Id: Ie4cea81c12a1a6f4935bfab57fb8eec2a898003e