undertaker-checkpatch: Properly find patched file
When a commit has a non-standard length hunk hash (more than 7
characters) in the "index $hash..$hash $mode" line, whatthepatch
will show the path to the changed file _including_ the leading
"a/", e.g. "a/path/to/file" instead of "path/to/file".

If this has happened, i.e. the file name starts with "a/",
strip it away.

Change-Id: Ib482fed3a322a78a7819137b6d8e0723070a4345