undertaker-checkpatch: Process line insertions from diff correctly
In commit d15e32c0f0c9da1a58ea857d27324ce65c834eec, we changed the
code in order to collect updates for processed lines in a new
variable, but compare to the old range of a block.

This, however, is only partially correct: When lines are removed,
the old line numbers from an unprocessed file are used as the
beginning of a hunk. On the other hand, for insertions the updated
range with all changes up to the current hunk are used.

This means, we need to compare the block range with the old range
when a deletion is processed, but with the updated range when lines
are inserted in a patch.

Change-Id: I10a629044bf5a18a17cf88dd1c3d6072db3f95b6