undertaker/blockconf: don't print config when blockpc isn't solvable
When "undertaker -j blockconf" was called on a dead block, a
configuration was printed. This configuration only contained the
transitive dependencies of the always_on items within a model, plus
the file preconditions. This might lead to unexpected behaviour for
the user, therefore we only print a warning that the block isn't
solvable.

Change-Id: I0c3731dcf7f7d1a349c7e95614ff1e6fcb01a5f7