undertaker: append the arch specific defect-types to the defect report
When we check for (global) defects, we analyze all blocks with every
loaded model when we have found a defect on the main architecture.
It's possible to have a different defect classifications for different
architectures. To be able to identify the problem faster, it might be
helpful to know, which architectures did have wich defect-type.
Therefore, we collect the "architecture -> defect-type"-mapping and
append it to the defect report.

Change-Id: Iae4d4410e74975f8868f64f8a57b13b066e3b86b