undertaker-kconfigdump: simplify dumpconf call
In kconfig, the order of the instructions is relevant, but we piped
the output of dumpconf through 'sort -u'. If we change rsf2model to
handle directives differently, we might have a invalid order.
However, up to now, we didn't gain much from sorting, so we are safe to
remove it.

Additionally, rsf2model now directly deals with comment-lines
(starting with '#'), so we don't need to filter these lines with
'grep -v' in undertaker-kconfigdump.

Change-Id: I72f7a74ca2057390ebf6c9f970ec4c098639d67b