satyr/kconfigdump: cleanup setting of environment variables
There is only one usage of the SRCARCH environment variable in linux
and this usage is in the root Kconfigfile, directly in the root
folder of a linux directory. However, we don't use this Kconfigfile
since 'um' is a special case and doesn't have the primary Kconfigfile
in 'arch/um'.

Otherwise, the SUBARCH environment variable has no impact on the logic
formulae and is always set to be the same as ARCH either way, so we
can simplify the logic in undertaker-kconfigdump.

Change-Id: I622c2af5b317d395bd6743dd03a56d41e3e09d39