undertaker-checkpatch: More checking on file paths
To be able to correctly work with whatthepatch 0.0.3, parse_patch()
needs to be adapted. In whatthepatch 0.0.2, when a file was moved or
newly created, the leading 'a/' or 'b/' before the actual path were
always present so they could be stripped away.

In 0.0.3, the parsing has changed and subsequently removes these
prefixes itself. We now check if the prefix is in the path and only
strip the first two characters if it really is 'a/' or 'b/',
respectively.

Change-Id: I8a4b0fbac1ebbcc6be0cfa989c39b10db991219d