rsf2model: handle CONFIG_{y,n} in rsf2model instead kconfigdump
This will ensure a more consistent model generation.

CONFIG_n was created, when a kconfig-option has no prompt and a
default n without any other dependencies. CONFIG_n should have been
either always_off or not contained in the conditions at all and not a
standard symbol.

CONFIG_y however, only is created in a special case and will now only
be added when this special case occurs, until we handle that case
differently.

Change-Id: Ibb2adaca1a5ec41a2098622de484b7140cc7368c