rsf2model: handle comparisons with non-existent items correctly
If we have an equality-comparison with an item that doesn't exist,
kconfig will never evaluate it as 'true'. Therefore, we can directly
add a contradiction in our model.
Example:
- for "A depends on (B=C)" we create an entry in our model
    A "((CONFIG_B && CONFIG_C) || (!CONFIG_B && !CONFIG_C))"
- if 'C' is not defined in kconfig, we will append "!CONFIG_C" in our
  missing analysis on a file that makes use of 'CONFIG_A'
- A is therefore reachable when CONFIG_B is not defined, which is a
  false negative. In kconfig, A would never be reachable.

The contradiction is generated by adding
"CONFIG_COMPARE_WITH_NONEXISTENT" to the always_off items.
The comparison is rewritten with the non-existent item:
(CONFIG_NON_EXISTENT && CONFIG_COMPARE_WITH_NONEXISTENT)

Thus, the identification of the problem is simple:
When the MUS-Formula contains a contradiction with
CONFIG_COMPARE_WITH_NONEXISTENT, look at the whole formula and search
for this symbol to find the offending symbol(s).

If we have a non-equality-comparison with an item that doesn't exist,
kconfig will always evaluate it as 'true'. Therefore, we can directly
add a tautology in our model. The tautology is generated with the
insertion of a unique __FREE__-Symbol that's always possible to
enable.

Change-Id: I8adc7929d4e978dbf22f1a4118b765603ef1ee69