python: simplify project detection
Currently, the code in python/vamos/golem/kbuild.py tries to get the
specific version from git or the Makefile of a particular project to
determine what we're running on.

As this information is only used in logging messages, it can be
simplified to use the same logic as undertaker-kconfigdump does.

Change-Id: If260e6d42612decd272eeecadd26a8cc71639033