undertaker-checkpatch: apply patch to index
Without applying the patch to the local Git index, the changes of patch
will only be propagated to the files but not to the Git index.  Hence,
undertaker-checkpatch gives some false or empty reports when checking
the referential integrity of Kconfig options (i.e., it git greps the
tree).

This patch fixes the aforementioned issue by adding '--index' when
applying the specified patch.  I detected this behavior by checking the
Linux commit 04ddc5b5ad4e ("xtensa: xtfpga: provide defconfig with audio
subsystem") which adds a defconfig file with two undefined Kconfig
options that both ended up in empty reports.

Change-Id: I1e6ebfb786ff160a1b65bb2a8b3201c94bcb65c6