pyMinigolem: parse Kbuild files in Python
This change adds a Python implementation of a parser for Kbuild
Makefiles.

Specific implementations for different phases of the parsing
process can be specified through the "-c" parameter, after which
one must specify a directory. Inside this directory, all .py
files are searched for classes inheriting from any of

kbuildparse.base_classes.{InitClass, BeforePass, DuringPass,
        AfterPass, BeforeExit}.

These classes are instantiated and gathered in different lists and
their process() function is called at the respective point.

If "-c" is not provided, the parser defaults to Linux.

Change-Id: I5b49decfc00541f7cfc3bf47c4271a7e9fb04881