undertaker-checkpatch: Do not miss symbols when looking for transitive items
Currently, only the first line of the output of
"undertaker -j interesting" was evaluated when
block.get_transitive_items() tried to find everything related to the
current block. However, when self.ref_items has more than one item in
it, the undertaker will return the interesting set for every item on
a separate line.

This change evaluates every line independently and updates the return
value as needed.

Change-Id: I7d054e02997b12554453a42fb2c72f32beea389c