undertaker-checkpatch: Set the correct main model for defect analysis
Currently, the main model always defaults to 'x86' when a directory
is passed to the -m parameter. For files inside other architecture's
arch/ directories, this leads to misleading report output (e.g.,
commit 37655163ce1a in linux-next).

This change makes undertaker-checkpatch check if the current srcfile
is in an arch/ directory and set the mainmodel and architecture
parameter correctly.

Change-Id: I9f5acf6f97f6abccd5b6e57d49438bf5c8cf0b11