rsf2model: speed up get_type function
When get_type was called, we iterated over a list of items in the
database to search for a specific item and returned the symbol of that
item.
However, we can speed up this lookup if we collect the Items first in
a memoized dict and have a O(1) look up for each get_type call.

Change-Id: I8ee0e5853e00b4b04deca41776ccdd401c6fec4d