undertaker-checkpatch: report only globally missing symbols
In commit 2b266b7b515f ("undertaker-checkpatch: Set the correct main
model for defect analysis"), we changed undertaker-checkpatch to
switch to the architecture-specific main model when a file inside the
arch/ subdirectory is analyzed. Additionally, we set the "arch"
variable (which defaults to "", if the "-a" command line parameter was
not given).

This, however, slightly changes the behaviour of
defect_analysis.check_missing_defect(): The items returned by
block.get_transitive_items(mainmodel) are only looked up in the main
model, as opposed to all models. While this is the correct behaviour,
a lot of noise is generated and renders the output nearly useless.
Example: When we encounter a variable in the block.ref_items which
references an architecture, the generated report will list all other
(29 in v4.0) architectures as "referenced but not defined".

However, with this change, undertaker-checkpatch can only identify
globally missing symbols in kconfig. We will alter this behaviour in
another patchset in the future which will be based on the MUS output.

This change reverts the usage of the "arch" parameter for
check_missing_defect() to its old state, while leaving the switch to
the correct main model in place.

Additionally, the identification of the required main model is more
robust, and a warning message is emitted when the main model can not
be changed.

Change-Id: Ia037a297df8372a819b8494745a89270b21318ab