dumpconf: fix choice handling
In "drivers/usb/gadget/Kconfig" is a source statement in a "choice ...
endchoice" group. In the sourced Kconfig-file are a number of choice
items and another "choice declaration". Pasted together the
declarations look like this:

choice
  tristate "USB Gadget Drivers"

config USB_CONFIGFS
    tristate "..."

config USB_CONFIGFS_ACM
    bool "..."
    depends on USB_CONFIGFS

...

config USB_G_DBGP
    tristate "..."

choice
  prompt "EHCI Debug Device mode"
  depends on USB_G_DBGP

config printk
    bool "..."

endchoice

config USB_G_WEBCAM
    tristate "..."

...

endchoice

dumpconf had multiple problems with this declaration:

- USB_G_WEBCAM and following items weren't recognized as ChoiceItem
  since the choice ended after USB_CONFIGFS_ACM due to the structure
  of the kconfig internal list.
  - USB_CONFIGFS_ACM has a dependency on USB_CONFIGFS. Thus,
    USB_CONFIGFS_ACM is a indirect successor of USB_CONFIGFS (via
    menu->list) and NOT a choice value! After this indirect successor
    list has been processed, we searched for the parent and finished
    the choice.

- The first choice was represented with CHOICE_1 and the second with
  CHOICE_2.
  Due to our way to set CHOICE_x symbols, USB_G_WEBCAM and the
  following items got a dependency on CHOICE_2, while CHOICE_2 had a
  dependency on the ChoiceItem USB_G_DBGP. If they were recognized as
  ChoiceItems, they would have been dead.

- If CHOICE_2 wouldn't have been present, USB_G_DBGP, USB_G_WEBCAM and
  following, still would have a dependency on CHOICE_1. However, they
  were not recognized as ChoiceItems. Therefore they were not present
  in the depencendy of CHOICE_1 and we would have needed another
  ChoiceItem to be enabled to satisfy the conditions for CHOICE_1.
  This would lead us to generate invalid configurations.

======================================================================

Additionally, this change fixes another problem.
When a symbol has a dependency on m (i.e. depends on m), kconfig adds
a automatic_symbol (flag SYMBOL_AUTO) to the dependencies that _might_
indicate, CONFIG_MODULES has to be defined. However, the symbol
doesn't point to CONFIG_MODULES, so we simply ignore it.

Our previous way to handle this symbol was to print
'"CHOICE_%d" choice_count' when a symbol didn't have a name.
That lead us to have dependencies on the last visited choice
structure when a symbol had a dependency on m.

Change-Id: Id3e4d81e278f9ee246d4224a506ba321ea4858e1