undertaker/symbolpc: don't print transitive conditions if missing
When the undertaker was asked for a symbol precondition, it always
printed the transitive conditions of the model (everything that's
always_on with their dependencies), regardless if the symbol was
contained in the model or not.
That might be heavily misleading for users and not helpful.
Thus, we now print an error message to indicate that item was missing
in the main model and exit before printing anything.

Change-Id: I7e39ea12b309876f35618dbead52cb6ac9c736bc