undertaker: use 'locally' in the filename of defect reports
The undertaker finds a lot of local defects. Up to now, we used the
main-arch in the filename of the defect report, if the defect was only
found on a few but not all architectures.
To have 'x86' in the filename suggests, local deads only occurs on
'x86', while it's a lot more diverse.
With this change, we only use 'locally' in the filename for local
deads.

Change-Id: I9fedebddc6fc2899626532fbd1c5b33b7f47522d