rsf2model: fix not-equal comparison translation for non-tristates
In commit 060a6d5c59 we changed the translation for equality
comparisons to include '_MODULE' symbols only when both symbols are
tristate.
This commit does a similar translation change for not-equals checks.

As the other commit, this change has no functional impact on the
undertaker, undertaker-checkpatch would report the _MODULE-Symbols as
missing, which might confuse users.

Change-Id: Ifc8a9fa5c6255ff496c2d62900cba3521335780c