undertaker-checkpatch: don't grep binary files
undertaker-checkpatch breaks in such cases, where a specified pattern
matches binary files, where the grep output for this file is for instance:
    Binary file drivers/staging/ft1000/ft1000-pcmcia/ft1000.img matches

The regular expressions to extract the file name and the corresponding
line don't match on this which caused undertaker-checkpatch to break.

This change fixes it by passing '-I' to git grep to ignore binary files.

Change-Id: If47de248b15612ef66c01651ae2e59d15524c58e