undertaker-checkpatch: only grep bare option strings in Kconfig space
To check the referential integrity of Kconfig, undertaker-checkpatch
checks if a newly referenced Kconfig option is undefined, or if a
removed Kconfig option is still referenced somewhere.  In such case, we
'git grep' this option in the tree.

The current implementation of grepping such missing options can lead to
considerable output sizes as the regex searches for the option with
CONFIG_ prefix, with _MODULE suffix and the bare string without suffix
or prefix.  However, a Kconfig option only carries the aforementioned
prefix and suffix in Make and CPP syntax.

This commit changes the behavior to only grep bare option strings in
Kconfig itself to avoid a to verbose (and partially wrong output).

Change-Id: Icb86b66c6d9b5fd668db9584357daddb8e05b8da