Explicitely set CFLAGS to "" when building picosat
On systems like Gentoo or Archlinux, a user can specify their own
system-wide CFLAGS for packaging. However, picosat is quite picky
when it comes to the right combination of preprocessor and compiler
flags, so we want it to use its own combination.

Take care of this by setting CFLAGS to the empty string when
calling the picosat Makefile from the toplevel Makefile.

Change-Id: If4b8f9104ab5f8e198e7b8a432240c3ed924f166