undertaker-checkpatch: add option '--commit'
Add git commit support to undertaker-checkpatch.  By passing the option
-c/--commit to undertaker-checkpatch, it checks the specified commit
instead of a patch file.

Note that this option relies on resetting the HEAD of the current git
repository, so that undertaker-checkpatch warns in case the HEAD is
dirty.  Hence, add another option '--force' to explicitly ignore the
warning of a dirty HEAD and continue execution.

Change-Id: Idfcfd7945f5027228b1a7c8c380e3ab2c45fbcb2