undertaker-checkpatch: similar symbols for missing Kconfig symbols
For each referential integrity violation (i.e., missing defects or newly
added but undefined symbols in Kconfig) add a line of 'Similiar
symbols:' to each report.  Note that a maximum of three similar symbols
be will printed.

The same code (based on difflib) has been used in checkkconfigsymbols.py
for a while now and proved to be quite helpful.

 --- Example of a missing defect:

New defect: arch/microblaze/mm/init.c:B8:307:309:missing.globally.dead
    CONFIG_HIGHMEM_START_BOOL is referenced but not defined in Kconfig

    Similar symbols: CONFIG_HIGHMEM, CONFIG_TIMER_STATS, CONFIG_IBM_RTL

 --- Example of a referential integrity violation:

X68_64 is not defined in Kconfig but newly referenced in:
    arch/x86/Kconfig:773:    depends on X68_64

    Similar symbols: X86_64, X86, M686

Change-Id: Id67da143dcc0087f08a8bde3ce6b3a9bc9f4e9b2