undertaker: Fix parentheses for multi-condition #elif
Currently, undertaker does not correctly parenthesize its
formulas when dealing with #elif blocks which have a
condition consisting of multiple alternatives.

As an example, if we have the following structure:

 #if defined(CONFIG_A)
 // B0
 #elif defined(CONFIG_B) || defined(CONFIG_C)
 // B1
 #endif

up to now, the formula generated for B1 by undertaker was
B1
&& ( B1 <-> (CONFIG_B) || (CONFIG_C) && ( ! (B0) ) )
&& ( B0 <-> (CONFIG_A) )
&& B00

As the precedence of "&&" is higher than the one of "||",
the SAT solver could generate a valid formula with CONFIG_B
and CONFIG_A enabled which is solvable but does not
correctly represent the conditions from the code - to reach
B1, CONFIG_A always has to be disabled (#elif for B1).

Fix this by putting additional brackets around the condition
for the currently checked block. Also add the new parentheses
to other testcases where necessary.

Change-Id: If2ea18650abba5b9ddd04998dc3e9e488358b93e