undertaker-checkpatch: Make it aware of kbuild defects
When a kbuild defect is found, we print the file precondition for
this file. As there could be different formulae depending on the
architecture, the output is grouped by architecture. For defects in
a particular block, the block's presence condition is also printed.

As a contradiction might not necessarily be in this information
alone but could also be in the transitive dependencies, a note is
printed to make the user aware of further possibilities.

Change-Id: Ic5528cc418a270cad1cc5383fcfb155998843437