undertaker-checkpatch: ignore internal always off symbol
For kconfig defects, we print symbols that are marked as always off
and always on to hint to the user which symbols _might_ be responsible
for the defect. However, we introduced an undertaker internal always
off symbol CONFIG_COMPARE_WITH_NONEXISTENT which might confuse
users when checkpatch prints it as hint. Therefore we ignore this
symbol in the output.

Change-Id: I407c056d8beefe426d8eea19cfb9e9b1cd8a9543