Predator: Fix compilation with AspectC++ 1.2
Copy adjusted "buildString" function to build a string from a unit
from Puma sources.  In aspectc++-1.2, this got moved around (and
refactored!) and the goal is to be able to compile against many
versions of Puma. This function is rather simple and we want exactly
this behavior, so copying is an acceptable approach.

Change-Id: I5f99070503bb5328ca077bc74fe4a44f6311cd9d