vampyr:Configuration.py: don't raise exception for differing items
All assignments of an expanded configuration are compared with the
initial configuration.  Don't raise an exception when assignments differ
since our models are not a 100% accurate mapping of the Kconfig
constraints.

During evaluation we encountered cases where Kconfig changes the values
of some items to fit the default values (e.g., defaults of choices).
Although the expanded config differs, the new assignment may in fact be
correct and should not raise an exception.  Such differing assignments
are now indicated as warnings, leaving the choice of validity to the
users.

Change-Id: Ia590441bdfce976eb3e8f6114c80ce38add6b897